+++ tests/parsing.rs
@@ -31,8 +31,7 @@ fn test_package_segment_boundaries() -> Result<(), Box<dyn std::error::Error>> {
     #[cfg(feature = "signature-meta")]
     {
         use rpm::signature::pgp::Signer;
-        let signing_key = common::keys::v4::rsa_private();
-        let signer = Signer::load_from_asc_bytes(signing_key.as_ref())?;
+        let signer = Signer::load_from_asc_file(common::keys::v4::RSA_4K_PRIVATE)?;
         let constructed_pkg_with_sig =
             rpm::PackageBuilder::new("empty-package", "0", "MIT", "x86_64", "")
                 .build_and_sign(signer)?;